Watchdog added.
I had a problem with the modem needing lots of power and my power supply couldn't handle the spike of currants and the modem just turned off. But not all ways it turned off and just because this device is meant to be always turned on I added a watchdog.