Big Update (untested)
I added a web interface kind of thing.
You can change the mac address of the PC you want to turn on and the phone number you want to get messages on.
Otherwise the "turn PC on " and "shut down PC" buttons do nothing for now. I will add functionality later.
For now after reboot the Arduino will reset the mac address and the phone number to the default one.